pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strepsuis-phylotrait"
version = "1.0.0"
description = "Phylogenetic clustering and binary trait analysis for bacterial genomics"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MK-vet", email = "support@strepsuis-suite.org"}
]
keywords = [
    "bioinformatics",
    "genomics",
    "antimicrobial-resistance",
    "bacterial-genomics",
    "streptococcus-suis"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: OS Independent"
]

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "statsmodels>=0.13.0",
    "jinja2>=3.0.0",
    "openpyxl>=3.0.0",
    "xlsxwriter>=3.0.0",
    "kaleido>=0.2.0",
    "tqdm>=4.62.0",
    "joblib>=1.1.0",
    "psutil>=5.8.0",
    "biopython>=1.79",
    "networkx>=2.6.0",
    "prince>=0.7.0",
    "umap-learn>=0.5.0",
    "optuna>=3.0.0",
    "mlxtend>=0.21.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
    "pdoc3>=0.10.0",
    "flake8>=4.0.0"
]

[project.urls]
Homepage = "https://github.com/MK-vet/strepsuis-phylotrait"
Repository = "https://github.com/MK-vet/strepsuis-phylotrait.git"
Issues = "https://github.com/MK-vet/strepsuis-phylotrait/issues"

[project.scripts]
strepsuis-phylotrait = "strepsuis_phylotrait.cli:main"

[tool.setuptools]
packages = ["strepsuis_phylotrait"]
